requirements.txt
elevenlabs
python-dotenv
fastrtc
loguru
langchain-community
langchain-ollama
langgraph
fastrtc[vad]
soundfile
litellm
groq
espeakng-loader
phonemizer-fork
# LlamaIndex dependencies
llama-index
llama-index-llms-litellm
google-auth

